Remove duplication of relationships in Neo4j
We were creating two relationships for every node relationship in
most cases (e.g. 'OWNED_BY' and 'OWNER'). This is bad practice
because it decreases performance, and the relationships can become
out of sync. We can achieve the same results by altering the
relationship direction in the Cypher queries.

For more information, read:
https://graphaware.com/neo4j/2013/10/11/neo4j-bidirectional-relationships.html