Refactor the dist-git scraper to not reuse processes while multi-processing
It turns out that using a multiprocess.Pool will reuse the processes rather than
destroy it and create a new one. This makes it so that the memory leak in one of
the libraries is persisting. This commit refactors this so that processes are
terminated after each chunk and created for every chunk processed.